refactor metamask snap connection logic
